Better refresh strategy for settings screen.
The settings screen is only refresh when it is entered. This
means that the user can edit the forms, without suddenly
changing values. A potential improvement is to update the
interface even if the tabs are switched (but they don't have
a onEnter function yet).

The current auto-failover status on the servers screens
is updated whenever the node is polled for server information.

Change-Id: Ib0404777f0f8a2647d3c7cf731cb5547f1362b95
Reviewed-on: http://review.couchbase.org/7010
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>