always unsubscribe bucket details renders before drawing buckets
If buckets list cell becomes undefined and we turn entire container
into big spinner, we still keep old subscriptions for bucket settings
details.

This is suboptimal because keeps a bunch of bucket details cells alive
as well as their associations with (now removed, but alive) DOM
elements.

But most importantly it's causing funny self-feedback loop on
IOCenter.staleness when logout is performed when network is lost.

Here's what's happening:

* IOCenter error queue contains a bunch of requests

* we do logout, clear few 'root' cells thus cancelling those requests

* IOCenter.status now becomes healthy (because no bad requests are
  known)

* Some bucket detail cell 'interested' is subscribed to
  IOCenter.staleness and gets recomputed

* This 'interested' cell becomes true and we recompute main details
  cell. Sending bucket details request

* This request fails turning IOCenter.staleness back to true

* This makes 'interested' cell turn to false, which makes us 'detach'
  from main details cell againg, cancelling request and turning
  IOCenter.staleness back to false

All this eats 100% CPU so a bit embarrasing.

Change-Id: I0c32ba8efefb65f911d654af47f8c61129c32711
Reviewed-on: http://review.couchbase.org/9982
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Benjamin Young <benjamin@couchbase.com>