use "Starting Up" string for undefined replication status
Right after POST-ing replication document we don't have any status due
to replication manager needing some time to notice it and start
replication. In this period of time replication status is
undefined. Thus we report undefined as "Starting Up".

There may be some other reasons for undefined status, but those should
be nearly impossible in practice, so doing what we do is fine.

Change-Id: I13c94ca57bad2b7dd2549549119b68767b278d38
Reviewed-on: http://review.couchbase.org/11138
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>