MB-6167: fixed saving map/reduce functions into wrong views
This was caused by idiotic behavior of lazy Cells that I introduced
some time ago. See test case that's now fails.

Basicaly I wanted re-subscribtion to detached cells to not recompute
formulas and not re-fetch values (if formula computes to future). And
that's causing current code to keep old value when cell is
detached. And that in turn caused #getValue to just use old value
instead of triggering cell recomputation.

Because fixing it either requires bigger changes or change of behavior
it's a bit too scary for now. Thus I added failing test and changed
affected cells to eager type.

Change-Id: I9c181d89e62df2dbd8ed9fe692b1a6cbbf51b574
Reviewed-on: http://review.couchbase.org/19514
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>