MB-6776 Per bucket setting to enable/disable flush.
During bucket creation or update 'flush_enabled' flag can be provided
to specify if flush should be enabled for the bucket. By default it's
disabled. When flush is disabled for some bucket then flush URI is not
advertised in bucket details. On the UI we will not show flush button
in such a case.

Change-Id: I459f6b0db32b824334382aade3c23539e47ad51d
Reviewed-on: http://review.couchbase.org/21318
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>