MB-7223: MB-7124: correctly fix sample document id capitalization
Some time ago unfortunately some css classes were reused in some new
parts of 2.0 essentially encouranging regression seen in
MB-7223.

Basically if in one place you want capitalization and another you
dont, you can 'fix' capitalization by changing CSS rule in one place
simultaneously braking it in other. That's because CSS classes are not
semantical and have unpredictable effects on unpredictable places.

So right fix of MB-7124 is to do it in a way that we started doing
once we realized scope of CSS mess we're in. We're mostly using inline
styles now.

Change-Id: I267a7698a499e35b96dd49e3672f4cbf60d9d66e
Reviewed-on: http://review.couchbase.org/22694
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>