MB-100: nuked long unused _replicator_infos ddoc and related APIs
Prior switching to task-based replications listing API we had that
original "_replicator database as API" idea that we largely inheritted
from couchdb. And part of that was special design doc that aggregated
status of replications out of info docs that were mutually replicated
between _replicator dbs of all nodes. Long before 2.0 GA we stopped
replicating info docs and prior to 2.0 GA we reimplemented XDCR
replication listing and status reporting on top of tasks. But
apparently remains of old approach were not killed. Now is time.

Change-Id: I8cad4d70eaef371314d2a2968acec9c6dd53db54
Reviewed-on: http://review.couchbase.org/25749
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>