MB-11596: implemented cheap and dirty way to disable alerts
We still have some reports of alerts preventing folks from using
UI. This is not how alerts are supposed to work but so far we're
unable to reproduce this and it's unclear what exactly happens in
those deployments.

In order to at least allow folks with this problems to proceed we're
implementing hidden feature that disables alerts completely. This
happens while ?disablePoorMansAlerts is appended to
index.html. Similarly to internal settings dialog. So it doesn't
disable alerts permanently on in specific browser tab that has
?disablePoorMansAlerts added to url.

Change-Id: If5b86c70a22c34506f137af36fab54344e881716
Reviewed-on: http://review.couchbase.org/40074
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>