MB-12388: Show build number next to release version in Admin UI
This is a simple UI change to show the version of the software that's running a little differently.
Previously we used to show the version (in "full mode") as:
     2.5.1 Enterprise Edition (build-1083-rel)
Now we show:
     2.5.1-1083 Enterprise Edition (build-1083-rel)

The idea of this work is that Support and Customers will treat 2.5.1-1083 as the full descriptor
of the release the customer is running, which is clearly critical when we are shipping a customer a
fix as part of a maintenance pack or even a one-off.

Change-Id: I1ac792ff8c3c531080d40e8adefaeaf77b2d8660
Reviewed-on: http://review.couchbase.org/42670
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>