MB-12222: update notifications section
In this implementation the launchpad will be called each
time when we call reloadState. This is not exactly
how the launchpad working in cell based implementation, but
ability to set the descendant depth of the reload would be
very useful for the lunchpad functionality and for mnApp in general.
Next release of ui-router has all chance to contain what we need.
https://github.com/angular-ui/ui-router/issues/1612.

Change-Id: If25a8f98dfe5aaea1be7eaff5b398802fb6d122c
Reviewed-on: http://review.couchbase.org/46940
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>