Add pluggable UI frontend framework
We need to add a little "framework" on the frontend to allow pluggable UI
components registers themselves with the UI platform and let us know where
and how they should be rendered.

This change proposes this initial API.

The key ideas are: flexibility, robustness and simplicity. Pluggable
components register themselves via a config object rather than very specific
API calls (for instance like addAdminTab() or things of this nature.)
This gives us loose coupling in that even if we add support later for things
like allowing the UI component to let us know in which position to add the
tab, the API doesn't change. We also promise to ignore attributes of the
config object that we don't understand which maximizes our chances to be
backwards compatible in the face of a changing set of config attributes
that we do support.

Change-Id: I8cc1a8ae39b76a2f92e2c79b680c9b4dc9685abe
Reviewed-on: http://review.couchbase.org/57176
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>