Change index.frag.html -> head.frag.html
The name is starting to bug me as these fragments aren't
generalized arbitrary fragments of what would normally go in
an HTML file. Rather, these fragments will be injected into
the <head> element of the main index.html file of our
frontend app -- so I figure let's pick a name that describes
what these things are and not what they're not.

Change-Id: I32672d9bb30b6bb1f907d93fa241f88d0a2f7f3d
Reviewed-on: http://review.couchbase.org/57220
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Anders Nygren <anders.nygren@couchbase.com>