MB-17295: allow pluggable tabs to specify preceding OOTB tab
Prior to this fix, pluggable tabs could only plug in at the end of the
row of tabs. This changes now treats all tabs on the tab bar (whether
out of the box or pluggable) uniformly and additionally allows
pluggable tabs to specify which tab they are after (their preceding tab.)

This will allow the query UI tab to specify an after value of
'admin.data_buckets' and get placed in the desired position on the
tab bar.

Change-Id: I12dd99251070a50e6fd73a8974d2653d4648cd85
Reviewed-on: http://review.couchbase.org/59446
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>