MB-17262: correct "show this server" link
The "Show this Server" link is shown on the specific state analytics
page. The title of the graph is computed as the IP address of the
host with the port removed if it's 8091. This is done by this bit
of code:
https://github.com/couchbase/ns_server/blob/master/src/menelaus_stats.erl#L2394

What this means is that the title is an incorrect hostname to pass
as the statsHostname parameter when the link is clicked.

What should be passed is the name.slice(1) as is done in the old UI.

Change-Id: I3dd6f00daad09122a980e4e042c48a1190c6d054
Reviewed-on: http://review.couchbase.org/59777
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>