MB-18289: support including lib/*.min.js files in index.min.html
And add angular.min.js to take advantage of this capability.
Angular.js weighs in at around 1 MB while its minified version
is 145 KB, which is a pretty significant saving in terms of
first UI page load.

Logic during creation of index.min.html is as follows:
if lib/{library}.js is encountered in index.html and
there exists a file lib/{library}.min.js
then this file is assumed to be a minified version of the
library and is included in place of {library}.js in the
generated index.min.html.

Angular.min.js is a pretty aggressively minified library.
However, it looks to preserve the object property names just
fine. E.g. the Promise and Deferred constructors are renamed
(safe because it's private) but the name of the $$state
member property is retained because clients can reach inside
the promise and access this (even though they are not supposed
to.)

Change-Id: Ica324b284f93287221c7446cd79445744cc560b0
Reviewed-on: http://review.couchbase.org/61973
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>