spock-ui: update look for initialization wizard
The wizard doesn't have the step 6 screen anymore, where the user had
the opportunity to adjust RAM quotas for the services. Instead of that UI
re-uses the serversMemoryQuotaDialog. It allows us to keep the related
code in one place and to follow the DRY rules without loosing of any
functionality.

Change-Id: I21ae6751d5d6c8314e127cdc7fa6f78dee0b8dfe
Reviewed-on: http://review.couchbase.org/67167
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>