Change information collection to comply with GDPR
With Couchbase's new GDPR policies, we need to be very careful in
handling any person data we collect. We had a GDPR meeting this week
concerning phone home data, and we came to the following conclusions:

1) We should not collect Name/Email/Company Name, because we have
never actually used that information, and we don't want to have to
come up with a privacy statement and figure out how to safely handle
all the data.

2) We need some kind of privacy statement in the configuration page
relating to "Enable software update notifications in the web console."

To solve #1, I removed the "Register for Updates" section in the
configuration panel. To solve #2, I copied the toolip that currently
exists in the "Settings" panel to the "New Cluster / Configure" panel.

I also removed replication IP address from the XDCR information
included in PhoneHome, because that is potentially identifiable
information.

Change-Id: Id6e983a3cc9474852a5672b515aff88da66467c9
Reviewed-on: http://review.couchbase.org/92205
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Rob Ashcom <rob.ashcom@gmail.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>