MB-29616 - Can't drop replica indexes from UI.
The REST endpoint /indexStatus returns multiple entries for replicated
indexes, with the replicas having names like "#primary (replica
1)". That is not a valid index name, so it can't be dropped.

It turns out that indexes are not permitted to have spaces in their
names. This fix hides the "Drop" index button for indexes with spaces
in their names, meaning that they are replicas.

Change-Id: I6ffd4e3eb04dfd5e1a07faf0a528ed17ec8427a3
Reviewed-on: http://review.couchbase.org/94314
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>