change scenarios schema and adapt UI to it
{
   version: 393221, //associated with ns_server version
   scenarios: [{
        groups: [ids....]
        other keys....
   }...],
   groups: [{
        charts: [ids....]
        other keys....
   }...],
   charts: [{
        other keys....
   }...]
}

The goal is to make data migration as easy as possible.
When we want to upgrade schema we go through all target
records and assign new keys/values but do not remove the
old keys/values. In this case all records will be backward
compatible. So we will have hybrid version of the json
that works for UI with different versions. In case we
drop support for some version we can remove the
undesirable fields during upgrade.

Example of how to perform upgrade:

if (profile.version < poolDefault.versions["70"]) {
   //change profile
   profile.version = poolDefault.versions["70"];
   //save profile
}

Also, this change introduces mnStoreService that helps
work with stored data.

And also, after saving the chat, UI doesn't reload the
entire charts list anymore, but only the chat that
has been saved. This significantly improves the user
experience.

Change-Id: I6ebf8125e7ac1600ae7e6a1dbcbd0f8618b5f260
Reviewed-on: http://review.couchbase.org/113509
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>